+++ src/Concise/Console/Command.php
@@ -7,6 +7,7 @@ use Concise\Console\ResultPrinter\DefaultResultPrinter;
 use Concise\Console\ResultPrinter\ResultPrinterProxy;
 use Concise\Console\TestRunner\DefaultTestRunner;
 use Concise\Console\Theme\DefaultTheme;
+use Concise\Extensions\Pho\PhoTestRunner;
 use Exception;
 use PHPUnit_TextUI_Command;
 use PHPUnit_TextUI_TestRunner;
@@ -31,7 +32,13 @@ class Command extends PHPUnit_TextUI_Command
         ) {
             $resultPrinter->setVerbose(true);
         }
-        $testRunner = new DefaultTestRunner($this->arguments['loader']);
+
+        if ($this->phoExtensionIsInstalled()) {
+            $testRunner = new PhoTestRunner($this->arguments['loader']);
+        } else {
+            $testRunner = new DefaultTestRunner($this->arguments['loader']);
+        }
+
         $testRunner->setPrinter(new ResultPrinterProxy($resultPrinter));
 
         return $testRunner;
@@ -121,5 +128,13 @@ class Command extends PHPUnit_TextUI_Command
                     break;
             }
         }
+    }
+
+    /**
+     * @return bool
+     */
+    protected function phoExtensionIsInstalled()
+    {
+        return class_exists('pho\Runnable\Spec');
     }
 }
